project view on work page
